+++ bin/read.js
@@ -19,7 +19,10 @@ runMain(async () => {
         process.exit(2)
     }
 
-    const knx = await KnxLink.connect(gateway)
+    const knx = await KnxLink.connect(gateway, {
+        onError: error => console.error(error.message),
+        onCemiFrame: () => {}
+    })
 
     const shutdown = async (code = 0) => {
         await knx.disconnect()