+++ src/lib/connection/KnxLink/KnxConnection/connect/index.ts
@@ -1,2 +1,2 @@
-export type { Connect, KnxLinkOptions, InternalLinkInfo, LinkInfo } from './types'
+export type { Connect, RequiredKnxLinkOptions, InternalLinkInfo, LinkInfo } from './types'
 export { default } from './connect'