+++ src/lib/connection/KnxLink/index.ts
@@ -1,4 +1,15 @@
-export { KnxLinkOptions } from './types'
-export type { LinkInfo } from './types'
+export type {
+    KnxLinkOptions,
+    RequiredKnxLinkOptions,
+    KnxLinkConnectOptions,
+    OnError,
+    KnxDisconnectedReason,
+    KnxConnectingEvent,
+    KnxReconnectingEvent,
+    KnxDisconnectedEvent,
+    LinkInfo
+} from './types'
+export type { KnxEventMap, KnxEventName, KnxEventListener } from './KnxEventEmitter'
 export { KnxLink } from './KnxLink'
+export { KnxEventEmitter } from './KnxEventEmitter'
 export { type DatapointConstructor, type KnxGroupSchema } from './types'