+++ src/lib/dpts/f32.spec.ts
@@ -1,10 +1,9 @@
-import EventEmitter from 'events'
-import { KnxLink, KnxLinkOptions } from '../connection'
+import { KnxLink, RequiredKnxLinkOptions } from '../connection'
 import { DPT_Value_Electric_Current, DPT_Value_Frequency, DPT_Value_Power, DPT_Value_Power_Factor } from './f32'
 
-function createF32<T>(DataType: new (address: string, link: KnxLink, options: KnxLinkOptions) => T): T {
+function createF32<T>(DataType: new (address: string, link: KnxLink, options: RequiredKnxLinkOptions) => T): T {
     const link = { sendCemiFrame: jest.fn().mockResolvedValue(undefined) } as unknown as KnxLink
-    const options = { events: new EventEmitter(), readTimeout: 1000 } as KnxLinkOptions
+    const options = { readTimeout: 1000 } as RequiredKnxLinkOptions
 
     return new DataType('3/1/1', link, options)
 }